django_inventory/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2023.4
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
